+++ src/main/java/io/fabric8/maven/docker/BuildMojo.java
@@ -1,8 +1,5 @@
 package io.fabric8.maven.docker;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.util.Date;
 import java.util.List;
 
@@ -11,8 +8,6 @@ import io.fabric8.maven.docker.util.ImageName;
 import io.fabric8.maven.docker.access.DockerAccessException;
 import io.fabric8.maven.docker.config.BuildImageConfiguration;
 import io.fabric8.maven.docker.service.ServiceHub;
-import io.fabric8.maven.docker.util.ImageNameFormatter;
-import org.apache.commons.io.IOUtils;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
@@ -25,27 +20,33 @@ import io.fabric8.maven.docker.util.EnvUtil;
  *
  * @author roland
  * @since 28.07.14
- *
  */
 @Mojo(name = "build", defaultPhase = LifecyclePhase.INSTALL)
 public class BuildMojo extends AbstractBuildSupportMojo {
 
-    @Parameter(property="docker.skipTags", defaultValue="false")
-    private boolean skipTags;
+    @Parameter(property = "docker.skip.tag", defaultValue = "false")
+    private boolean skipTag;
+
+    @Parameter(property = "docker.skip.build", defaultValue = "false")
+    private boolean skipBuild;
 
     @Override
     protected void executeInternal(ServiceHub hub) throws DockerAccessException, MojoExecutionException {
+        if (skipBuild) {
+            return;
+        }
         for (ImageConfiguration imageConfig : getResolvedImages()) {
             BuildImageConfiguration buildConfig = imageConfig.getBuildConfiguration();
 
             if (buildConfig != null) {
                 if (buildConfig.skip()) {
-                    log.info("%s : Skipped building",imageConfig.getDescription());
+                    log.info("%s : Skipped building", imageConfig.getDescription());
                 } else {
                     buildAndTag(hub, imageConfig);
                 }
             }
         }
+
     }
 
     // We ignore an already existing date file and always return the current date
@@ -57,17 +58,17 @@ public class BuildMojo extends AbstractBuildSupportMojo {
     private void buildAndTag(ServiceHub hub, ImageConfiguration imageConfig)
         throws MojoExecutionException, DockerAccessException {
         buildImage(hub, imageConfig);
-        if (!skipTags) {
+        if (!skipTag) {
             tagImage(imageConfig.getName(), imageConfig, hub.getDockerAccess());
         }
     }
-    
+
     private void tagImage(String imageName, ImageConfiguration imageConfig, DockerAccess dockerAccess)
-            throws DockerAccessException, MojoExecutionException {
+        throws DockerAccessException, MojoExecutionException {
 
         List<String> tags = imageConfig.getBuildConfiguration().getTags();
         if (tags.size() > 0) {
-            log.info("%s: Tag with %s",imageConfig.getDescription(),EnvUtil.stringJoin(tags, ","));
+            log.info("%s: Tag with %s", imageConfig.getDescription(), EnvUtil.stringJoin(tags, ","));
 
             for (String tag : tags) {
                 if (tag != null) {